pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = ""
authors = ["abdulrahman <abdelr7manfared@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
ipython = "^8.5.0"
django-model-utils = "^4.2.0"
django-bootstrap-datepicker-plus = "^4.0.0"
Pillow = "^9.2.0"
django-imagekit = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
